+++ src/main/java/br/com/fittank/fittank/model/Gerente.java
@@ -19,7 +19,7 @@ public class Gerente implements Serializable {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Long id;
+	private Integer id;
 	
 	@OneToOne
 	@JoinColumn(name = "fk_usuario", nullable = false)
@@ -29,16 +29,16 @@ public class Gerente implements Serializable {
 		
 	}
 	
-	public Gerente(Long id, Usuario usuario) {
+	public Gerente(Integer id, Usuario usuario) {
 		this.id = id;
 		this.usuario = usuario;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
